python-requirements.txt
appdirs==1.4.3
bcrypt==3.1.3
bson==0.4.7
cffi==1.10.0
falcon==1.2.0
falcon-auth==1.0.3
gunicorn==19.7.1
mysqlclient==1.3.10
packaging==16.8
pkg-resources==0.0.0
py==1.4.33
pycparser==2.17
PyJWT==1.5.0
pymongo==3.4.0
pytest==3.1.0
python-mimeparse==1.6.0
pytz==2017.2
six==1.10.0
SQLAlchemy==1.1.11
